requirements.txt
asgiref==3.5.0
cachetools==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.4
Django==4.0.3
django-environ==0.8.1
djangorestframework==3.13.1
filelock==3.6.0
google-api-core==2.10.0
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
oauth2client==4.1.3
Pillow==9.1.0
platformdirs==2.5.2
protobuf==4.21.5
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.1
requests==2.27.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.9
xmltodict==0.13.0
